# manager/api/api-spec.bundled.yaml
openapi: 3.0.0
info:
  version: 0.0.0
  title: MaEVe CSMS
  description: Internal API to interact with the MaEVe CSMS, external clients should use OCPI.
  contact:
    name: MaEVe team
    email: maeve-team@thoughtworks.com
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:9410/api/v0
    description: The local development server
paths:
...
  /cs/{csId}/transactions:
    get:
      summary: List transactions for a charge station
      description: |
        Lists transactions for the specified charge station with optional filtering
        by status, date range, and pagination.
      operationId: listTransactions
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          required: true
          schema:
            type: string
            maxLength: 28
        - name: status
          in: query
          description: 'Filter by transaction status (default: all)'
          required: false
          schema:
            type: string
            enum:
              - active
              - completed
              - all
            default: all
        - name: limit
          in: query
          description: 'Maximum number of transactions to return (default: 50, max: 100)'
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: 'Number of transactions to skip for pagination (default: 0)'
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: startDate
          in: query
          description: Filter transactions starting on or after this date (ISO8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter transactions starting on or before this date (ISO8601)
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/transaction/{transactionId}:
    get:
      summary: Get transaction details
      description: |
        Retrieves detailed information about a specific transaction including
        all meter values.
      operationId: getTransactionDetails
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          required: true
          schema:
            type: string
            maxLength: 28
        - name: transactionId
          in: path
          description: The transaction identifier
          required: true
          schema:
            type: string
            maxLength: 36
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetail'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /cs/{csId}/start-transaction:
    post:
      summary: Remote start transaction
      description: |
        Remotely starts a charging transaction on the specified charge station.
        This is an async operation that returns immediately with 202 Accepted.
      operationId: remoteStartTransaction
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteStartTransactionRequest'
      responses:
        '202':
          description: Accepted - remote start initiated
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/stop-transaction:
    post:
      summary: Remote stop transaction
      description: |
        Remotely stops an active charging transaction on the specified charge station.
        This is an async operation that returns immediately with 202 Accepted.
      operationId: remoteStopTransaction
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteStopTransactionRequest'
      responses:
        '202':
          description: Accepted - remote stop initiated
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  schemas:
...
    RemoteStartTransactionRequest:
      type: object
      description: Request to remotely start a charging transaction
      required:
        - idTag
      properties:
        idTag:
          type: string
          maxLength: 20
          description: The identifier that should be used to authorize the transaction
        connectorId:
          type: integer
          minimum: 1
          description: Optional connector ID to start charging on
        chargingProfile:
          $ref: '#/components/schemas/ChargingProfile'
    RemoteStopTransactionRequest:
      type: object
      description: Request to remotely stop a charging transaction
      required:
        - transactionId
      properties:
        transactionId:
          type: string
          description: The identifier of the transaction to stop
    Status:
      type: object
      description: HTTP status
      required:
        - status
      properties:
        status:
          type: string
          description: The status description
        error:
          type: string
          description: The error details
...
    TransactionList:
      type: object
      description: List of transactions with pagination
      required:
        - transactions
        - total
        - limit
        - offset
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionSummary'
        total:
          type: integer
          format: int64
          description: Total number of transactions matching the filter
        limit:
          type: integer
          description: Maximum number of results returned
        offset:
          type: integer
          description: Number of results skipped
    TransactionSummary:
      type: object
      description: Summary of a transaction
      required:
        - transactionId
        - idTag
        - startTime
        - meterStart
        - status
      properties:
        transactionId:
          type: string
          description: The transaction identifier
        idTag:
          type: string
          description: The identifier used to start the transaction
        startTime:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction started
        stopTime:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction stopped (null if active)
        meterStart:
          type: integer
          description: Meter value in Wh at start of transaction
        meterStop:
          type: integer
          description: Meter value in Wh at end of transaction (null if active)
        status:
          type: string
          enum:
            - active
            - completed
          description: Status of the transaction
    TransactionDetail:
      type: object
      description: Detailed transaction information including meter values
      required:
        - transactionId
        - idTag
        - startTime
        - meterStart
        - status
        - meterValues
      properties:
        transactionId:
          type: string
          description: The transaction identifier
        idTag:
          type: string
          description: The identifier used to start the transaction
        tokenType:
          type: string
          description: The type of token used
        startTime:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction started
        stopTime:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the transaction stopped (null if active)
        meterStart:
          type: integer
          description: Meter value in Wh at start of transaction
        meterStop:
          type: integer
          description: Meter value in Wh at end of transaction (null if active)
        status:
          type: string
          enum:
            - active
            - completed
          description: Status of the transaction
        meterValues:
          type: array
          items:
            $ref: '#/components/schemas/MeterValue'
          description: All meter values recorded during the transaction
...
    MeterValuesSampledValue:
      type: object
      description: A single sampled value in a meter reading
      required:
        - value
      properties:
        value:
          type: string
          description: Measured value (numeric as string)
        context:
          type: string
          description: Reading context (e.g., Sample.Periodic, Transaction.Begin)
        format:
          type: string
          description: Value format (Raw or SignedData)
        measurand:
          type: string
          description: Type of measurement (e.g., Energy.Active.Import.Register, Power.Active.Import)
        phase:
          type: string
          description: Phase measured (L1, L2, L3, N, L1-N, L2-N, L3-N, L1-L2, L2-L3, L3-L1)
        location:
          type: string
          description: Location of measurement (Cable, EV, Inlet, Outlet, Body)
        unit:
          type: string
          description: Unit of measurement (Wh, kWh, varh, kvarh, W, kW, VA, kVA, var, kvar, A, V, K, Celsius, Fahrenheit, Percent)
    MeterValue:
      type: object
      description: A set of meter values at a specific timestamp
      required:
        - timestamp
        - sampledValue
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the meter reading
        connectorId:
          type: integer
          description: Connector ID for the reading
        evseId:
          type: integer
          description: EVSE ID for OCPP 2.0.1
        transactionId:
          type: string
          description: Transaction ID (if part of a transaction)
        sampledValue:
          type: array
          items:
            $ref: '#/components/schemas/MeterValuesSampledValue'
          description: Array of sampled values
...
    ChargingProfile:
      type: object
      required:
        - chargingProfileId
        - stackLevel
        - chargingProfilePurpose
        - chargingProfileKind
        - chargingSchedule
      properties:
        chargingProfileId:
          type: integer
          format: int32
        transactionId:
          type: integer
          format: int32
        stackLevel:
          type: integer
          format: int32
          minimum: 0
        chargingProfilePurpose:
          type: string
          enum:
            - TxProfile
            - TxDefaultProfile
            - ChargePointMaxProfile
        chargingProfileKind:
          type: string
          enum:
            - Absolute
            - Relative
            - Recurring
        recurrencyKind:
          type: string
          enum:
            - Daily
            - Weekly
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        chargingSchedule:
          $ref: '#/components/schemas/ChargingSchedule'
    ChargingSchedule:
      type: object
      required:
        - chargingRateUnit
        - chargingSchedulePeriod
      properties:
        duration:
          type: integer
          format: int32
        startSchedule:
          type: string
          format: date-time
        chargingRateUnit:
          type: string
          enum:
            - W
            - A
        chargingSchedulePeriod:
          type: array
          items:
            $ref: '#/components/schemas/ChargingSchedulePeriod'
          minItems: 1
        minChargingRate:
          type: number
          format: double
    ChargingSchedulePeriod:
      type: object
      required:
        - startPeriod
        - limit
      properties:
        startPeriod:
          type: integer
          format: int32
          minimum: 0
        limit:
          type: number
          format: double
        numberPhases:
          type: integer
          format: int32
          minimum: 1
...